refactor: separate interfaces, add IMessage and ISendMesage, add SocketEvents
