+++ src/utils/enums/SocketEvent.ts
@@ -6,4 +6,6 @@ export enum SocketEvent {
   FinishGame = 'finishGame',
   UserVote = 'userVote',
   CreateGame = 'createGame',
+  MessageCreated = 'message:created',
+  MessageCreate = 'message:create'
 }
