+++ src/utils/interfaces/IMessage.ts
@@ -1,0 +1,5 @@
+export interface IMessage {
+  userName: string,
+  message: string,
+  id: number
+}
