+++ src/utils/interfaces/ISendMessage.ts
@@ -1,0 +1,5 @@
+export interface ISendMessage {
+  text: string;
+  roomId: string;
+  userId: number
+}
