+++ src/utils/interfaces/IUser.ts
