+++ src/utils/interfaces/index.ts
@@ -1,1 +1,3 @@
-export type { IUser } from './interfaces';
+export type { IUser } from './IUser';
+export type { IMessage } from './IMessage';
+export type { ISendMessage } from './ISendMessage';
